products_api/serializers.py
from rest_framework import serializers
from .models import *


class ProductSerializer(serializers.ModelSerializer):

    user = serializers.HiddenField(default=serializers.CurrentUserDefault())

    class Meta:
        model = Product
        fields = '__all__'

